pyproject.toml
[build-system]
requires = [
    "nanobind >= 1.8.0",
    "petsc4py",
    "scikit-build-core[pyproject]"
]
build-backend = "scikit_build_core.build"

[project]
name = "RBniCSx"
version = "0.0.dev1"
authors = [
    {name = "Francesco Ballarin", email = "francesco.ballarin@unicatt.it"},
]
maintainers = [
    {name = "Francesco Ballarin", email = "francesco.ballarin@unicatt.it"},
]
description = "Reduced order modelling in FEniCSx"
license = {file = "COPYING"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: GNU Lesser General Public License v3 or later (LGPLv3+)",
    "Operating System :: POSIX",
    "Operating System :: POSIX :: Linux",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Mathematics",
    "Topic :: Software Development :: Libraries :: Python Modules"
]
dependencies = [
    "mpi4py",
    "numpy >= 1.21.0",
    "petsc4py",
    "plum-dispatch",
    "slepc4py",
    "typing-extensions; python_version < '3.11'"
]

[project.urls]
homepage = "https://www.rbnicsproject.org"
repository = "https://github.com/RBniCS/RBniCSx"
issues = "https://github.com/RBniCS/RBniCSx/issues"
funding = "https://github.com/sponsors/francesco-ballarin"

[project.optional-dependencies]
backends = [
    "adios4dolfinx @ git+https://github.com/jorgensd/adios4dolfinx.git",
    "fenics-dolfinx >=0.8.0.dev0, <0.9.0",
    "multiphenicsx @ git+https://github.com/multiphenics/multiphenicsx.git",
    "ufl4rom @ git+https://github.com/RBniCS/ufl4rom.git"
]
docs = [
    "sphinx"
]
lint = [
    "clang-format",
    "cmakelang",
    "isort",
    "mypy",
    "nbqa",
    "ruff",
    "yamllint"
]
tests = [
    "coverage[toml]",
    "nbvalx[unit_tests]",
    "pytest",
    "scipy"
]
tutorials = [
    "gmsh",
    "nbvalx[notebooks]",
    "plotly",
    "rbnicsx[backends]",
    "viskex @ git+https://github.com/viskex/viskex.git"
]

[tool.isort]
line_length = 120
multi_line_output = 4
order_by_type = false
combine_as_imports = true
known_first_party = ["rbnicsx"]

[tool.mypy]
check_untyped_defs = true
disallow_any_unimported = true
disallow_untyped_defs = true
implicit_reexport = true
no_implicit_optional = true
pretty = true
show_error_codes = true
strict = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    # "dolfinx",
    # "dolfinx.*",
    "gmsh",
    "ipyparallel",
    "petsc4py",
    "petsc4py.PETSc",
    "plotly",
    "plotly.*",
    "plum",
    "scipy",
    "scipy.*",
    "slepc4py",
    "slepc4py.SLEPc",
    "ufl"
]
ignore_missing_imports = true

[tool.pytest.ini_options]

[tool.ruff]
line-length = 120
select = ["ANN", "D", "E", "F", "FLY", "ICN", "N", "NPY", "Q", "RUF", "UP", "W"]
ignore = ["ANN101"]

[tool.ruff.per-file-ignores]
"rbnicsx/**/__init__.py" = ["F401"]
"rbnicsx/**/*.py" = ["N802", "N803", "N806"]
"tests/unit/**/*.py" = ["N802", "N803", "N806"]
"tutorials/0*.py" = ["D100", "F401", "N803", "N806"]
"tutorials/1*.py" = ["D100", "F401", "N803", "N806"]
"tutorials/**/0*.py" = ["D100", "F401", "N803", "N806"]
"tutorials/**/1*.py" = ["D100", "F401", "N803", "N806"]

[tool.ruff.pycodestyle]
max-doc-length = 120

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.scikit-build]
cmake.source-dir = "rbnicsx/_cpp"
wheel.packages = ["rbnicsx"]
sdist.exclude = ["*.cpp", "*.h", "CMakeLists.txt"]
